refactor: finish cleanup backlog and archive Zig-era plans
Split WDBX format + CLI modules, share WAL Mutation serde, extract agent
util/os and complete arg parsing, add MCP malformed-bearer contracts and a
Rust-rescoped #647 audit note, and move completed Zig plans to archive.
