+++ crates/abi-cli/src/wdbx/db.rs
@@ -1,0 +1,163 @@
+//! WDBX `db` subcommand: segment checkpoints, WAL recovery, and snapshot integrity.
+//!
+//! Split from the flat `wdbx` CLI module; dispatch lives in `super::run`.
+
+use crate::app::Outcome;
+use crate::usage::is_help_token;
+use crate::wdbx::paths_from_cli_base;
+use abi_wdbx::{Snapshot, Wal};
+use std::fmt::Write;
+
+pub(crate) const DB_HELP: &str = "usage: abi wdbx db <init|verify|compact> <path> [keep]\n\nManage segment checkpoints, WAL recovery, and snapshot integrity.\n";
+
+fn init_db(raw_path: &str) -> Outcome {
+    let paths = match paths_from_cli_base(raw_path) {
+        Ok(paths) => paths,
+        Err(detail) => return super::error("db init failed", detail),
+    };
+    if let Err(detail) = std::fs::create_dir_all(&paths.dir) {
+        return super::error("db init failed", detail);
+    }
+    if let Err(detail) = abi_wdbx::segments::reset(&paths) {
+        return super::error("db init failed", detail);
+    }
+    if let Err(detail) = std::fs::remove_file(paths.index())
+        && detail.kind() != std::io::ErrorKind::NotFound
+    {
+        return super::error("db init failed", detail);
+    }
+    if let Err(detail) = std::fs::remove_file(paths.mirror_epoch())
+        && detail.kind() != std::io::ErrorKind::NotFound
+    {
+        return super::error("db init failed", detail);
+    }
+    let wal = abi_wdbx::wal::wal_path(&paths);
+    if let Err(detail) = std::fs::remove_file(&wal)
+        && detail.kind() != std::io::ErrorKind::NotFound
+    {
+        return super::error("db init failed", detail);
+    }
+    if let Err(detail) = abi_wdbx::persistence::flush(&paths, &Snapshot::new()) {
+        return super::error("db init failed", detail);
+    }
+    Outcome::stderr(
+        format!("initialized empty WDBX segment checkpoint at {raw_path}\n"),
+        0,
+    )
+}
+
+fn verify_db(raw_path: &str) -> Outcome {
+    let paths = match paths_from_cli_base(raw_path) {
+        Ok(paths) => paths,
+        Err(detail) => return super::error("verify FAILED", detail),
+    };
+    let (snapshot, checkpoint_source) = match abi_wdbx::store::load_checkpoint_with_source(&paths) {
+        Ok(loaded) => loaded,
+        Err(detail) => {
+            return super::error(&format!("verify FAILED: checkpoint {raw_path}"), detail);
+        }
+    };
+    let chain_valid = snapshot.verify_chain_strict().is_ok();
+    let stats = snapshot.stats;
+    let source = match checkpoint_source {
+        abi_wdbx::store::CheckpointSource::Empty => "empty",
+        abi_wdbx::store::CheckpointSource::Legacy { .. } => "snapshot",
+        abi_wdbx::store::CheckpointSource::Segment { .. } => "segment",
+    };
+    let checkpoint_epoch = checkpoint_source.epoch();
+    let mut report = format!(
+        "checkpoint OK: source={source} epoch={checkpoint_epoch} kv={} vectors={} blocks={} spatial={} temporal_nodes={} temporal_edges={} chain_valid={chain_valid}\n",
+        stats.kv_entries,
+        stats.vectors,
+        stats.blocks,
+        stats.spatial_records,
+        stats.temporal_nodes,
+        stats.temporal_edges
+    );
+
+    let wal_path = abi_wdbx::wal::wal_path(&paths);
+    if !wal_path.is_file() {
+        return Outcome::stderr(report, u8::from(!chain_valid));
+    }
+    let wal = match Wal::read(&wal_path) {
+        Ok(wal) => wal,
+        Err(detail) => {
+            writeln!(
+                report,
+                "WAL verify FAILED: {}: {detail}",
+                wal_path.display()
+            )
+            .expect("writing to a String cannot fail");
+            return Outcome::stderr(report, 1);
+        }
+    };
+    if wal.base_epoch != checkpoint_epoch {
+        writeln!(
+            report,
+            "WAL note: frames={} base_epoch={} predates checkpoint epoch={checkpoint_epoch}; discarded on recovery",
+            wal.len(),
+            wal.base_epoch
+        )
+        .expect("writing to a String cannot fail");
+        return Outcome::stderr(report, u8::from(!chain_valid));
+    }
+
+    let mut merged = snapshot;
+    if let Err(detail) = wal.replay_onto(&mut merged) {
+        writeln!(
+            report,
+            "WAL replay FAILED: {}: {detail}",
+            wal_path.display()
+        )
+        .expect("writing to a String cannot fail");
+        return Outcome::stderr(report, 1);
+    }
+    let merged_valid = merged.verify_chain_strict().is_ok();
+    writeln!(
+        report,
+        "WAL OK: frames={} merged_blocks={} merged_chain_valid={merged_valid}",
+        wal.len(),
+        merged.blocks.len()
+    )
+    .expect("writing to a String cannot fail");
+    Outcome::stderr(report, u8::from(!(chain_valid && merged_valid)))
+}
+
+fn compact_db(raw_path: &str, keep_latest: usize) -> Outcome {
+    let paths = match paths_from_cli_base(raw_path) {
+        Ok(paths) => paths,
+        Err(detail) => return super::error("compact FAILED", detail),
+    };
+    let result = match abi_wdbx::segments::compact_retain_latest(&paths, keep_latest) {
+        Ok(result) => result,
+        Err(detail) => return super::error(&format!("compact FAILED: {raw_path}"), detail),
+    };
+    let latest = result
+        .latest_epoch
+        .map_or_else(|| "none".to_owned(), |epoch| epoch.to_string());
+    let mut report = format!(
+        "compacted WDBX segments: path={raw_path} keep_latest={} before={} after={} deleted={} latest_epoch={latest}",
+        result.keep_latest, result.before, result.after, result.deleted
+    );
+    if let Some(watermark) = result.watermark_epoch {
+        write!(report, " watermark_epoch={watermark}").expect("writing to a String cannot fail");
+    }
+    report.push('\n');
+    Outcome::stderr(report, 0)
+}
+
+pub(crate) fn run_db(args: &[String]) -> Outcome {
+    if args.len() == 1 && is_help_token(&args[0]) {
+        return Outcome::stderr(DB_HELP.to_owned(), 0);
+    }
+    match args {
+        [operation, path] if operation == "init" => init_db(path),
+        [operation, path] if operation == "verify" => verify_db(path),
+        [operation, path] if operation == "compact" => compact_db(path, 2),
+        [operation, path, keep] if operation == "compact" => match keep.parse::<usize>() {
+            Ok(keep) if keep > 0 => compact_db(path, keep),
+            _ => super::usage(),
+        },
+        _ => super::usage(),
+    }
+}
