+++ crates/abi-wdbx/src/lib.rs
@@ -40,17 +40,21 @@ pub mod durable;
 pub mod entropy;
 pub mod fhe;
 pub mod format;
+pub mod hash;
 pub mod hnsw;
 pub mod index;
+pub mod manifest;
 pub mod multiway;
 pub mod mvcc;
 pub mod net_line;
 pub mod neural_compress;
 pub mod persistence;
 pub mod rate_limit;
+pub mod record;
 pub mod remote_compute;
 pub mod rest;
 pub mod retrieval;
+pub mod segment;
 pub mod segments;
 pub mod spatial;
 pub mod store;
